% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.sport-package.R
\docType{package}
\name{rc.sport-package}
\alias{rc.sport}
\alias{rc.sport-package}
\title{rc.sport: Research Compendium of the GIPSO Biological Rhythms and Sport SQLR}
\description{
Research compendium of the GIPSO biological rhythms and sport systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/rc.sport/}
  \item \url{https://github.com/gipso/rc.sport/}
  \item Report bugs at \url{https://github.com/gipso/rc.sport/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rafael Salgueiro Poli \email{rafael.poli@usp.br} (\href{https://orcid.org/0000-0002-2017-1216}{ORCID}) [copyright holder]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
