% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport-package.R
\docType{package}
\name{sport-package}
\alias{sport}
\alias{sport-package}
\title{sport: R Package for the Biological Rhythms and Sports Project}
\description{
Research compendium of the biological rhythms and sports project: a systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/sport/}
  \item \url{https://github.com/gipso/sport/}
  \item Report bugs at \url{https://github.com/gipso/sport/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rafael Salgueiro Poli \email{rafael.poli@usp.br} (\href{https://orcid.org/0000-0002-2017-1216}{ORCID}) [copyright holder]
  \item Ana Amelia Benedito-Silva \email{aamelia@usp.br} (\href{https://orcid.org/0000-0003-4976-2623}{ORCID}) [scientific advisor]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
